Receipt_OCR_Experiments/requirements.txt
Flask==3.0.3
Pillow==10.0.1
pytesseract==0.3.10